Changed Game.isMobile to touch input detector
Instead of iOS/Android detector, for better Windows/alternate platform
compatibility. “isMobile” is a misnomer, and only left named as-is for
the sake of not breaking old implementations. Should be changed to
“hasTouch” (or similar) in future.

Also added Game.isAndroid property to complement Game.isIOS. We should
determine whether or not these properties are useful, and if so,
whether or not they are correctly implemented.